pyproject.toml
[tool.poetry]
name = "likers"
version = "0.2.0"
description = "Likers social network"
authors = ["Constantine <ceeveeya@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9"
Django = ">=2.2.3"
djangorestframework = ">=3.9.4"
djangorestframework-simplejwt = ">=4.6.0"
Faker = ">=2.0.1"
psycopg2 = ">=2.8.3"
requests = ">=2.22.0"
django-environ = ">=0.4.5"
celery = ">=5.0.5"
gunicorn = ">=20.0.4"
python-dotenv = ">=0.15.0"
drf-yasg = ">=1.20.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = ">=3.2"
black = "^23.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \venv
  | _build
  | build
  | dist
  | node_modules
  | .serverless
  | .circleci
  | .github
  | test-reports
  | .idea
)/
'''